cclient/serialization/json: refactor into multiple files
